changed as per PR suggestions
