--- src/MiddleWares/index.js
@@ -1,5 +1,0 @@
-import Logger from "./Logger.js";
-import errorHandler from "./errorHandler.js";
-import addCustomHeader from "./addCustomHeader.js";
-import rateLimit from "./rateLimit.js";
-export { Logger, errorHandler, addCustomHeader, rateLimit };
