--- src/controllers/OtherController.js
@@ -1,1 +1,0 @@
-
