+++ src/middleWares/index.js
@@ -1,0 +1,6 @@
+import logger from "./logger.js";
+import errorHandler from "./errorHandler.js";
+import addCustomHeader from "./addCustomHeader.js";
+import rateLimit from "./rateLimit.js";
+import authenticate from "./authenticate.js";
+export { logger, errorHandler, addCustomHeader, rateLimit, authenticate };
