--- src/routes/otherRoutes.js
@@ -1,1 +1,0 @@
-
