+++ src/routes/userRoutes.js
@@ -1,31 +1,25 @@
 import express from "express";
 import jwt from "jsonwebtoken";
-import { GetData, PostData } from "../controllers/UserController.js";
-import { addCustomHeader, rateLimit } from "../MiddleWares/index.js";
+import { GetData, PostData } from "../controllers/userController.js";
+import { addCustomHeader, authenticate } from "../middleWares/index.js";
 const router = express.Router();
-const secretKey = "myNameIsMunendraKumarKushwaha";
-const authenticate = (req, res, next) => {
-  const token = req.header("authorization");
-  if (!token) {
-    return res.status(401).json({ message: "unauthorized" });
-  }
-  try {
-    const decoded = jwt.verify(token, secretKey);
-    req.user = decoded;
-    console.log(req.user);
-    next();
-  } catch (error) {
-    return res.status(401).json({ message: "unauthorized" });
-  }
-};
 
 router.route("/").get(authenticate, addCustomHeader, GetData);
 router.route("/").post(authenticate, addCustomHeader, PostData);
+
+const secretKey = "myNameIsMunendraKumarKushwaha";
+
 router.post("/login", (req, res) => {
   const user = { id: 1, name: "monu" };
-  jwt.sign(user, secretKey, { expiresIn: "30m" }, (err, token) => {
-    res.json({ token });
-  });
+  try {
+    jwt.sign(user, secretKey, { expiresIn: "30m" }, (err, token) => {
+      res.status(200).json({ token });
+    });
+  } catch (error) {
+    error.message = "Error in generation of token";
+    err.status = 400;
+    next(error);
+  }
 });
 
 export default router;
